authentication

+++ base/views.py
@@ -2,13 +2,62 @@ from django.shortcuts import render, redirect
 from .models import Room, Topic
 from .forms import RoomForm
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate,login, logout
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
+from django.http import  HttpResponse
+from django.contrib.auth.forms import UserCreationForm
 
 # Create your views here.
+
+def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
+    if request.method == 'POST':
+        username = request.POST.get('username')
+        password = request.POST.get('password')
+        try:
+            user = User.objects.get(username=username)
+        except: 
+            print('except')
+            messages.error(request,'User does not exist')
+        user =  authenticate(request,username=username,password=password)
+        if user is not None:
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'username or password does not exist')
+    context={'page':page}
+    return render(request,'login_register.html',context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerPage(request):
+    page='register'
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user =  form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+    else:
+        messages.error(request,'Register failed')
+    context={'page':page,'form':form}
+    return render(request,'login_register.html',context)
+
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
     rooms =  Room.objects.filter(Q(topic__name__icontains=q) | Q(name__icontains=q) | Q(description__icontains=q))
     topics = Topic.objects.all()
-    context = {'rooms':rooms,'topics':topics}
+    room_count = rooms.count()
+    context = {'rooms':rooms,'topics':topics,'room_count':room_count}
     return render(request,'home.html',context)
 
 def room(request,pk):
@@ -16,6 +65,7 @@ def room(request,pk):
     context = {'room':room}
     return render(request,'room.html',context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -26,9 +76,12 @@ def createRoom(request):
     context = {'form':form}
     return render(request,'room_form.html',context)
 
+@login_required(login_url='login')
 def updateRoom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here')
     if request.method == 'POST':
         form = RoomForm(request.POST,instance=room)
         if form.is_valid():
@@ -37,8 +90,11 @@ def updateRoom(request,pk):
     context = {'form':form}
     return render(request,'room_form.html',context)
 
+@login_required(login_url='login')
 def deleteRoom(request,pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('Your are not allowed here')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
